+++ catalog/app/components/Preview/renderers/index.js
@@ -1,3 +1,4 @@
+export { default as Audio } from './Audio'
 export { default as DataFrame } from './DataFrame'
 export { default as ECharts } from './ECharts'
 export { default as Fcs } from './Fcs'